media: android_bug: fix play error when play again after alloc codec fail [1/1]
PD#SWPL-1701

Problem:
This an acodec bug, When alloc component fail, it will signel error,
and then reclaim the last component alloc a new component again.
But this time mFatalError alse is true. So when dequeueBufferFromNativeWindow
it will "not dequeuing from native window due to fatal error",
and return NULL. It duo to video can not play.

Solution:
This fixed is when alloc a new component again success, set mFatalError to false

Verify:
verified by p212

Change-Id: Ie0f9e7345b6235d08f2e65238c8a0ae8302a42f0
Signed-off-by: rongrong zhou <rongrong.zhou@amlogic.com>