cec: fix the fuzzy logic when cec is switched constantly [1/2]
PD#SWPL-15991

Problem:
When cec is switched to open and close constantly, the logic
in HdmiControlService is mixed up with enable and disable work.
The consequence is that more and more HdmiCecLocalDeviceTv
instances is living with HotplugDetection working, they are
constantly sending send physical address message to playbacks
while they can't really process the response.But the real tv
instance will be constantly excuting hanleReportPhysicalAddress
and then adding the device, then resulting in triggering the
onHdmiDeviceAdded callback with power status is off and excute
the deviceSelect work. This is why it feels more and more
stuck.

Solution:
Use synchronization to make sure the disable and enable work
is sperated in HdmiControlService.

Verify:
Marconi

Change-Id: I48deeca75532b7fafc166c2944979918f87df620
Signed-off-by: an.xi <an.xi@amlogic.com>