ViewRootImpl: remove reset mPreviousTransparentRegion [1/1]
PD#SWPL-22491

Problem:
transparent region error

Solution:
revert googlw's commit
commit 61a6e689188cd8743905e4f7b10440c7e50823f3
Author: Jack Palevich <jackpal@google.com>
Date: Fri Oct 9 17:37:50 2009 -0700

Prevent transparent region from getting out of sync with Surface Flinger

The ViewRoot is responsible for calculating the current view's surface's
transparent region and reporting this to the window server, which in
turn notifies Surface Flinger.

To minimize cross-process traffic, ViewRoot keeps track of the
most recent transparent region that it has reported to the window
server, and only sends updates when the region shape changes.

ViewRoot is effectively shadowing the internal state of the window
server, and usually this works fine, because ViewRoot is usually the
only thing that changes this value.

But when the surface is recreated, the window server resets the
transparent region to empty. So in order to properly update the window
server we need to reset our copy of its state to empty too.

Verify:
ok on newton.

Change-Id: I03eb654213f82f78d2e5102fb39a0fe03fe7eaff
Signed-off-by: Zhiwei Yan <zhiwei.yan@amlogic.com>