skia: Make Paint default to filtering Bitmaps [1/1]
PD#SWPL-26952

Problem:
Bug: 77949917
Test: I5737be68dccdc5ebb38291bf3515ae25553dcfc2

In P, the switch to Skia HW rendering lost a feature of HWUI's old
renderer: using a Paint with a BitmapShader resulted in filtering the
Bitmap, regardless of whether FILTER_BITMAP_FLAG was set.

Solution:
This somewhat restores that behavior, in that it results in filtering
BitmapShaders so long as the client doesn't explicitly turn it off. This
also means that it continues to be possible to get a pixellated drawing
if the client *does* explicitly turn it off.

Verify:
franklin

Change-Id: Idd08f4592a986209fa8fd2c1d725340be2896b96
Signed-off-by: binqi zhang <binqi.zhang@amlogic.com>